requirements.txt
djangorestframework==3.10.3
Django==2.2.5
coloredlogs==10.0
psycopg2==2.8.3
django-dotenv==1.4.2
django-heroku==0.3.1
gunicorn==19.9.0
django-crequest==2018.5.11
requests==2.21.0
pubchempy==1.0.4